requirements.txt
langchain 
PyPDF2 
streamlit 
openai 
faiss-cpu
tiktoken
python-dotenv